requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord.py==1.5.1
docopt==0.6.2
idna==2.10
multidict==4.7.6
pipreqs==0.4.10
requests==2.25.1
urllib3==1.26.2
yarg==0.1.9
yarl==1.5.1
